Big refactoring!
