+++ .gitignore
@@ -31,3 +31,44 @@ lichess_bot_refactored/.idea/vcs.xml
 lichess_bot_refactored/.idea/workspace.xml
 lichess_bot_refactored/.vscode/c_cpp_properties.json
 lichess_bot_refactored/.vscode/settings.json
+cmake-build-debug/Testing/Temporary/LastTest.log
+cmake-build-debug/CMakeFiles/TargetDirectories.txt
+cmake-build-debug/CMakeFiles/rules.ninja
+cmake-build-debug/CMakeFiles/3.22.3/CompilerIdC/a.out
+cmake-build-debug/CMakeFiles/3.22.3/CMakeDetermineCompilerABI_C.bin
+cmake-build-debug/CMakeFiles/3.22.3/CMakeCCompiler.cmake
+cmake-build-debug/cmake_install.cmake
+cmake-build-debug/build.ninja
+cmake-build-debug/.cmake/api/v1/reply/toolchains-v1-f90b60b2dfba8819cbe1.json
+cmake-build-debug/.cmake/api/v1/reply/target-Chess-Debug-d6b8417e2002ce29c715.json
+cmake-build-debug/.cmake/api/v1/reply/index-2022-06-13T03-45-15-0047.json
+cmake-build-debug/.cmake/api/v1/reply/directory-.-Debug-f5ebdc15457944623624.json
+cmake-build-debug/.cmake/api/v1/reply/codemodel-v2-c6622c44669d223815b6.json
+cmake-build-debug/.cmake/api/v1/reply/cmakeFiles-v1-de3e2b3a9662d7621bac.json
+cmake-build-debug/.cmake/api/v1/reply/cache-v2-1df2a8c2a01436cf03fb.json
+Chess
+*.cmake
+CMakeCache.txt
+CMakeFiles/3.23.2/CMakeDetermineCompilerABI_C.bin
+CMakeFiles/3.23.2/CompilerIdC/CMakeCCompilerId.c
+CMakeFiles/3.23.2/CompilerIdC/CMakeCCompilerId.o
+CMakeFiles/Chess.dir/build.make
+CMakeFiles/Chess.dir/compiler_depend.internal
+CMakeFiles/Chess.dir/compiler_depend.make
+CMakeFiles/Chess.dir/compiler_depend.ts
+CMakeFiles/Chess.dir/depend.make
+CMakeFiles/Chess.dir/flags.make
+CMakeFiles/Chess.dir/lichess_bot_refactored/ChessEngine.c.o
+CMakeFiles/Chess.dir/lichess_bot_refactored/ChessEngine.c.o.d
+CMakeFiles/Chess.dir/lichess_bot_refactored/lib/hdict.c.o.d
+CMakeFiles/Chess.dir/link.txt
+CMakeFiles/Chess.dir/progress.make
+CMakeFiles/cmake.check_cache
+CMakeFiles/CMakeError.log
+CMakeFiles/CMakeOutput.log
+CMakeFiles/Makefile2
+CMakeFiles/progress.marks
+CMakeFiles/TargetDirectories.txt
+lichess_bot_refactored/ChessEngine.c
+*.o
+*.d
