+++ lichess_bot_refactored/dataStructs.c
@@ -1,10 +1,15 @@
 #include <stdint.h>
 
+struct move_info {
+    char move[2];    // origin and destination squares (0 - 63)
+    char side;    // side moving white = 0 (whitePawns) or black = 7 (blackPawns)
+    char piece;    // see enum EPieceType
+};
+
 uint64_t not_a_file = 18374403900871474942ULL;
 uint64_t not_h_file = 9187201950435737471ULL;
 uint64_t not_hg_file = 4557430888798830399ULL;
 uint64_t not_ab_file = 18229723555195321596ULL;
-
 
 const int LS1Bindex64[64] = {
     0, 47,  1, 56, 48, 27,  2, 60,
@@ -15,28 +20,4 @@ const int LS1Bindex64[64] = {
     34, 51, 20, 43, 31, 22, 10, 45,
     25, 39, 14, 33, 19, 30,  9, 24,
     13, 18,  8, 12,  7,  6,  5, 63
-};
-
-struct move_info {
-    char move[2];    // origin and destination squares (0 - 63)
-    char side;    // side moving white = 0 (whitePawns) or black = 7 (blackPawns)
-    char piece;    // see enum EPieceType
-};
-
-// NOTE: Also have the mask_knight_attacks function. Delete one of them.
-const uint64_t knight_attack_lookup[64] = {
-    132096UL, 329728UL, 659712UL, 1319424UL, 2638848UL, 5277696UL, 10489856UL, 4202496UL, 33816580UL, 
-    84410376UL, 168886289UL, 337772578UL, 675545156UL, 1351090312UL, 2685403152UL, 1075839008UL, 
-    8657044482UL, 21609056261UL, 43234889994UL, 86469779988UL, 172939559976UL, 345879119952UL, 
-    687463207072UL, 275414786112UL, 2216203387392UL, 5531918402816UL, 11068131838464UL, 
-    22136263676928UL, 44272527353856UL, 88545054707712UL, 175990581010432UL, 70506185244672UL, 
-    567348067172352UL, 1416171111120896UL, 2833441750646784UL, 5666883501293568UL, 
-    11333767002587136UL, 22667534005174272UL, 45053588738670592UL, 18049583422636032UL, 
-    145241105196122112UL, 362539804446949376UL, 725361088165576704UL, 1450722176331153408UL, 
-    2901444352662306816UL, 5802888705324613632UL, 11533718717099671552UL, 
-    4620693356194824192UL, 288234782788157440UL, 576469569871282176UL, 1224997833292120064UL, 
-    2449995666584240128UL, 4899991333168480256UL, 9799982666336960512UL, 
-    1152939783987658752UL,2305878468463689728UL, 1128098930098176UL, 2257297371824128UL, 
-    4796069720358912UL, 9592139440717824UL, 19184278881435648UL, 38368557762871296UL, 
-    4679521487814656UL, 9077567998918656UL
 };
